src/dbt_to_cypher/extractor.py
"""
Module for extracting dependencies from dbt projects.
"""

from pathlib import Path
from typing import Dict, List, Optional

import yaml


class DbtDependencyExtractor:
    """
    Extract model and column-level dependencies from a dbt project.
    
    This class parses dbt manifest.json and other project files to build
    a comprehensive dependency graph including both model-level and 
    column-level lineage.
    """

    def __init__(self, project_path: Path):
        """
        Initialize the extractor with a dbt project path.
        
        Args:
            project_path: Path to the dbt project directory
        """
        self.project_path = Path(project_path)
        self.manifest_path = self.project_path / "target" / "manifest.json"
        self.catalog_path = self.project_path / "target" / "catalog.json"

    def extract_model_dependencies(self) -> Dict:
        """
        Extract model-level dependencies from the dbt project.
        
        Returns:
            Dictionary containing model dependency information
        """
        # TODO: Implement manifest.json parsing
        raise NotImplementedError("Model dependency extraction not yet implemented")

    def extract_column_dependencies(self) -> Dict:
        """
        Extract column-level dependencies from the dbt project.
        
        Returns:
            Dictionary containing column lineage information
        """
        # TODO: Implement column lineage extraction
        raise NotImplementedError("Column dependency extraction not yet implemented")

    def extract_all(self) -> Dict:
        """
        Extract both model and column-level dependencies.
        
        Returns:
            Complete dependency graph data structure
        """
        return {
            "models": self.extract_model_dependencies(),
            "columns": self.extract_column_dependencies(),
        }
